requirements.txt
streamlit>=1.30.0
anthropic>=0.18.0
python-docx>=1.1.0
markdown>=3.5.0
Pillow>=10.0.0
pandas>=2.0.0
plotly>=5.18.0
weasyprint>=60.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pytest>=7.4.0
